Fix an E_NOTICE

+++ WADF.php
@@ -772,15 +772,17 @@ class Tools_WADF {
 			foreach(explode(' ', $files) as $file) {
 				$force_remove = false;
 				$file = trim($file);
-				if ($file{0} == '+') { // force removal even if template doesn't exist
-					$file = substr($file, 1);
-					$force_remove = true;
-				}
-				chdir($deploy_path);
-				if (file_exists($file)) {
-					if ($force_remove || file_exists("$file.template")) {
-						$this->_debugOutput("  Removing $file", self::DEBUG_INFORMATION);
-						unlink($file);
+				if (strlen($file) > 0) {
+					if ($file{0} == '+') { // force removal even if template doesn't exist
+						$file = substr($file, 1);
+						$force_remove = true;
+					}
+					chdir($deploy_path);
+					if (file_exists($file)) {
+						if ($force_remove || file_exists("$file.template")) {
+							$this->_debugOutput("  Removing $file", self::DEBUG_INFORMATION);
+							unlink($file);
+						}
 					}
 				}
 			}
